refactor; added test command
